🔧 2021.3 Add WebP images and optimise page loading
The og:image meta property has a relative URL as its content. However, it does not seem to show up for Twitter and throws an error in the Open Graph debugger.
Therefore, I replaced the relative URL with an absolute URL on all pages for this property.

The PNG images are big and take time to load.
Therefore, I use the <picture> and <souce> tags to load the WebP images I generated. When not able to load, they will fall back to their PNG version.

The website always downloads custom fonts even if the user already has them locally. However, this is inefficient.
Therefore, I use the local() CSS function in the src descriptor to try and load local fonts before eventually downloading them if necessary.

While downloading the font files, the browser does not show the text.
Therefore, I change this behaviour with the font-display CSS descriptor. It will now appear with a fallback font and swap it with the downloaded font once available.