🐞 2021.4 Blur navigation bar's background
In normal circumstances, the secondary navigation bar is supposed to have a blurred background on browsers that support the backdrop-filter CSS property. However, for this property to take effect, the background colour should not be opaque, which it mistakenly was.
Therefore, I added a support query and changed the background colour to a new transparent one when the blur effect is supported.