🔧 2021.a.1 Adjust Christmas Guide media and texts
Who Am I? is being updated visually and renamed to Christmas Guide.
However, the website does not reflect the changes yet.
Therefore, I updated the assets and press kit to use the new icon and app screenshots. I have also removed mentions of the old name and changed the colours and paragraph texts to reflect the update.

The site aligns the navigation picker items on the apps page in the centre.
However, when using a notice on one of the items, their alignment seems incorrect.
Therefore, I changed the item alignment to be on the start.