🐞 2021.a.4 Fix privacy policy date
I had set Christmas Guide's privacy policy date to the 25th of September.
However, that date corresponded to the private beta release day and not the current public beta release day, which will fall on the same date as the website update.
Therefore, I modified the policy's date to the 13th of October, the currently scheduled date for the public beta release.