⭐️ 2021.9 Add new article and mention
I'm now ready to publish my newest Swift Tip, "Present share sheet in SwiftUI". Therefore, I prepared the page and assets for the article, and I linked it on the Articles and Swift Tip page.

The Code Completion podcast mentions Huh? in one of its episodes. Therefore, I added a link to the video on the Mentions page with its appropriate metadata and asset.