🔧 2021.10 Update featured Swift Tip and sitemap
The featured articles section on the home page should reflect the newly published Swift tip. Therefore, I replaced the old featured Swift Tip with the new one.

The sitemap should reflect the addition of a new page to the website and a change in page priorities. Therefore, I generated a new sitemap.