🐞 2022.4 Fix links and typos
Discovering that linking to a post with a final backslash opens an inexistent page, I removed the unwanted character where applicable.

Finding out that splitting a URL into multiple lines yields it dysfunctional, I removed all new lines within a link.

Willing to prevent typos associated with the asset name on the About page, I used the slugify filter and fixed a typo.