⭐️ 2022.14 Schedule a new article
⭐️ Feature

Planning articles for the coming months, I wrote and scheduled an article called How To Speak in a new category (Loud Letters).

🔧 Adjustment

Automating some actions of my workflow, I created a script and GitHub Workflow that uses Google cwebp tool to convert PNG images to WebP.

Seeing that the font size for lists was too small, I gave it the same font as for paragraphs and links in articles.

Wanting differentiation between article categories, I edited the colors for each article's hero image based on that category.